Use SIGKILL after delay and only if need
Sometimes OB is not reacting on SIGTERM, we need SIGKILL it.
This change add some delay before we do it.

+++ stop.sh
@@ -2,5 +2,10 @@
 for pid in `ps aux | grep "python.*openbazaar_daemon.py" | grep -v grep | awk '{print $2}'`; do
   echo "Sending SIGTERM to ${pid}"
   kill ${pid}
-  kill -9 ${pid}
+  sleep 10s
+  if ps -p $pid > /dev/null
+    then
+    echo "Still running, sending SIGKILL to ${pid}"
+    kill -9 ${pid}
+  fi
 done
